2 Guessing Game
Put the whole game in an infinite loop, convert the string to a u32, use
match to match for the 2 returns of the Result type, if Ok return the
number, if its Err continue to next iteration.
User Ordering::cmp trait to compare the 2 values, we are comparing the
secret number to the guess, and use another match to match for the
returns of type Ordering? Greater, Equal and Lessm then handle the logic
based on the returned value, if user guesses the number he wins and the
loop is broken.